backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.0.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configparser==7.2.0
Django==5.2.1
django-cors-headers==4.3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-yasg==1.21.7
ecdsa==0.19.1
freezegun==1.5.2
gunicorn==21.2.0
inflection==0.5.1
kombu==5.5.4
mongodb-migrations==1.3.1
mongoengine==0.29.1
mongomock==4.3.0
packaging==25.0
passlib==1.7.4
pillow==10.2.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
sentinels==1.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
